requirements.txt
gradio 
tensorflow 
matplotlib 
pillow